+++ lib/lib.go
@@ -20,7 +20,7 @@ import (
 )
 
 var client = &http.Client{
-	Timeout: 10 * time.Second, // Adjusted timeout to 10 seconds
+	Timeout: 1 * time.Second, // Adjusted timeout to 10 seconds
 	Transport: &http.Transport{
 		MaxIdleConns:        100,              // Increased maximum idle connections
 		MaxIdleConnsPerHost: 10,               // Increased maximum idle connections per host
