+++ nodes.toml
@@ -5,7 +5,7 @@ channel = "channel-1"
 denom = "uom"
 prefix = "mantra"
 gas_per_byte = 100
-base_gas = 41772713
+base_gas = 160000
 ibc_memo = "Contract store spam test"
 memo = "Storing compiled contract with randomized memo"
 ibc_memo_repeat = 10
@@ -14,6 +14,9 @@ rand_max = 30000
 revision_number = 4
 timeout_height = 21720608
 slip44 = 118
+
+broadcast_mode = "grpc" # or "rpc"
+
 
 msg_type = "bank_send"
 
@@ -42,3 +45,4 @@ precision = 3
 [nodes]
 rpc = ["http://127.0.0.1:26657"]
 api = "https://api.canary.mantrachain.dev:443"
+grpc = "localhost:9090"
